backend/summarySnap/requirements.txt
python-dotenv
selenium
beautifulsoup4
requests
Flask
coloredlogs
web3
appdirs
timeout_decorator
flask-cors
revChatGPT
